refactor: announcement api entities로 분리
PROF-86
