+++ src/entities/announcement/index.ts
@@ -1,0 +1,2 @@
+export { announcementQuries } from './api/query';
+export { type Announcement } from './api/schema';
